refactor(multiple): add utils to aria/private and consolidate sortDirectives usage
