+++ src/aria/private/utils/BUILD.bazel
@@ -1,0 +1,8 @@
+load("//tools:defaults.bzl", "ts_project")
+
+package(default_visibility = ["//visibility:public"])
+
+ts_project(
+    name = "utils",
+    srcs = ["element.ts"],
+)
