+++ src/aria/private/utils/element.ts
