+++ src/aria/tree/tree-item-group.ts
@@ -16,9 +16,8 @@ import {
   OnInit,
   OnDestroy,
 } from '@angular/core';
-import {TreeItemPattern, DeferredContent} from '../private';
+import {TreeItemPattern, DeferredContent, sortDirectives} from '../private';
 import type {TreeItem} from './tree-item';
-import {sortDirectives} from './utils';
 
 /**
  * Group that contains children tree items.
